refactor(AxesEditor)
